list_models.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()

def list_models():
    api_key = os.getenv("GOOGLE_API_KEY")
    if not api_key:
        print("GOOGLE_API_KEY not found.")
        return

    genai.configure(api_key=api_key)
    
    print("Listing ALL models...")
    try:
        for m in genai.list_models():
            print(f"Name: {m.name} | Methods: {m.supported_generation_methods}")
    except Exception as e:
        print(f"Error listing models: {e}")

if __name__ == "__main__":
    list_models()
