requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu
pypdf
streamlit
python-dotenv
google-generativeai
